Rename package to scalars
