+++ examples/client/server/src/main/kotlin/com/expediagroup/graphql/examples/model/ScalarWrapper.kt
@@ -1,7 +1,7 @@
 package com.expediagroup.graphql.examples.model
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.types.ID
+import com.expediagroup.graphql.scalars.ID
 import java.util.UUID
 
 @GraphQLDescription("Wrapper that holds all supported scalar types")
