+++ graphql-kotlin-schema-generator/src/test/kotlin/com/expediagroup/graphql/test/integration/NodeGraphTest.kt
@@ -19,7 +19,7 @@ package com.expediagroup.graphql.test.integration
 import com.expediagroup.graphql.TopLevelObject
 import com.expediagroup.graphql.testSchemaConfig
 import com.expediagroup.graphql.toSchema
-import com.expediagroup.graphql.types.ID
+import com.expediagroup.graphql.scalars.ID
 import graphql.schema.GraphQLObjectType
 import graphql.schema.GraphQLTypeReference
 import org.junit.jupiter.api.Test
